+++ api/core/v2/mutator_test.go
@@ -1,6 +1,7 @@
 package v2
 
 import (
+	"sort"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -29,4 +30,38 @@ func TestMutatorValidate(t *testing.T) {
 
 	// Valid mutator
 	assert.NoError(t, m.Validate())
+}
+
+func TestSortMutatorsByName(t *testing.T) {
+	a := FixtureMutator("Abernathy")
+	b := FixtureMutator("Bernard")
+	c := FixtureMutator("Clementine")
+	d := FixtureMutator("Dolores")
+
+	testCases := []struct {
+		name     string
+		inDir    bool
+		inChecks []*Mutator
+		expected []*Mutator
+	}{
+		{
+			name:     "Sorts ascending",
+			inDir:    true,
+			inChecks: []*Mutator{d, c, a, b},
+			expected: []*Mutator{a, b, c, d},
+		},
+		{
+			name:     "Sorts descending",
+			inDir:    false,
+			inChecks: []*Mutator{d, a, c, b},
+			expected: []*Mutator{d, c, b, a},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			sort.Sort(SortMutatorsByName(tc.inChecks, tc.inDir))
+			assert.EqualValues(t, tc.expected, tc.inChecks)
+		})
+	}
 }