+++ backend/apid/graphql/schema/mutator.gql.go
@@ -419,3 +419,453 @@ var _ObjectTypeMutatorDesc = graphql.ObjectDesc{
 		"toJSON":    _ObjTypeMutatorToJSONHandler,
 	},
 }
+
+// MutatorConnectionNodesFieldResolver implement to resolve requests for the MutatorConnection's nodes field.
+type MutatorConnectionNodesFieldResolver interface {
+	// Nodes implements response to request for nodes field.
+	Nodes(p graphql.ResolveParams) (interface{}, error)
+}
+
+// MutatorConnectionPageInfoFieldResolver implement to resolve requests for the MutatorConnection's pageInfo field.
+type MutatorConnectionPageInfoFieldResolver interface {
+	// PageInfo implements response to request for pageInfo field.
+	PageInfo(p graphql.ResolveParams) (interface{}, error)
+}
+
+//
+// MutatorConnectionFieldResolvers represents a collection of methods whose products represent the
+// response values of the 'MutatorConnection' type.
+//
+// == Example SDL
+//
+//   """
+//   Dog's are not hooman.
+//   """
+//   type Dog implements Pet {
+//     "name of this fine beast."
+//     name:  String!
+//
+//     "breed of this silly animal; probably shibe."
+//     breed: [Breed]
+//   }
+//
+// == Example generated interface
+//
+//   // DogResolver ...
+//   type DogFieldResolvers interface {
+//     DogNameFieldResolver
+//     DogBreedFieldResolver
+//
+//     // IsTypeOf is used to determine if a given value is associated with the Dog type
+//     IsTypeOf(interface{}, graphql.IsTypeOfParams) bool
+//   }
+//
+// == Example implementation ...
+//
+//   // DogResolver implements DogFieldResolvers interface
+//   type DogResolver struct {
+//     logger logrus.LogEntry
+//     store interface{
+//       store.BreedStore
+//       store.DogStore
+//     }
+//   }
+//
+//   // Name implements response to request for name field.
+//   func (r *DogResolver) Name(p graphql.ResolveParams) (interface{}, error) {
+//     // ... implementation details ...
+//     dog := p.Source.(DogGetter)
+//     return dog.GetName()
+//   }
+//
+//   // Breed implements response to request for breed field.
+//   func (r *DogResolver) Breed(p graphql.ResolveParams) (interface{}, error) {
+//     // ... implementation details ...
+//     dog := p.Source.(DogGetter)
+//     breed := r.store.GetBreed(dog.GetBreedName())
+//     return breed
+//   }
+//
+//   // IsTypeOf is used to determine if a given value is associated with the Dog type
+//   func (r *DogResolver) IsTypeOf(p graphql.IsTypeOfParams) bool {
+//     // ... implementation details ...
+//     _, ok := p.Value.(DogGetter)
+//     return ok
+//   }
+//
+type MutatorConnectionFieldResolvers interface {
+	MutatorConnectionNodesFieldResolver
+	MutatorConnectionPageInfoFieldResolver
+}
+
+// MutatorConnectionAliases implements all methods on MutatorConnectionFieldResolvers interface by using reflection to
+// match name of field to a field on the given value. Intent is reduce friction
+// of writing new resolvers by removing all the instances where you would simply
+// have the resolvers method return a field.
+//
+// == Example SDL
+//
+//    type Dog {
+//      name:   String!
+//      weight: Float!
+//      dob:    DateTime
+//      breed:  [Breed]
+//    }
+//
+// == Example generated aliases
+//
+//   type DogAliases struct {}
+//   func (_ DogAliases) Name(p graphql.ResolveParams) (interface{}, error) {
+//     // reflect...
+//   }
+//   func (_ DogAliases) Weight(p graphql.ResolveParams) (interface{}, error) {
+//     // reflect...
+//   }
+//   func (_ DogAliases) Dob(p graphql.ResolveParams) (interface{}, error) {
+//     // reflect...
+//   }
+//   func (_ DogAliases) Breed(p graphql.ResolveParams) (interface{}, error) {
+//     // reflect...
+//   }
+//
+// == Example Implementation
+//
+//   type DogResolver struct { // Implements DogResolver
+//     DogAliases
+//     store store.BreedStore
+//   }
+//
+//   // NOTE:
+//   // All other fields are satisified by DogAliases but since this one
+//   // requires hitting the store we implement it in our resolver.
+//   func (r *DogResolver) Breed(p graphql.ResolveParams) interface{} {
+//     dog := v.(*Dog)
+//     return r.BreedsById(dog.BreedIDs)
+//   }
+//
+type MutatorConnectionAliases struct{}
+
+// Nodes implements response to request for 'nodes' field.
+func (_ MutatorConnectionAliases) Nodes(p graphql.ResolveParams) (interface{}, error) {
+	val, err := graphql.DefaultResolver(p.Source, p.Info.FieldName)
+	return val, err
+}
+
+// PageInfo implements response to request for 'pageInfo' field.
+func (_ MutatorConnectionAliases) PageInfo(p graphql.ResolveParams) (interface{}, error) {
+	val, err := graphql.DefaultResolver(p.Source, p.Info.FieldName)
+	return val, err
+}
+
+// MutatorConnectionType A connection to a sequence of records.
+var MutatorConnectionType = graphql.NewType("MutatorConnection", graphql.ObjectKind)
+
+// RegisterMutatorConnection registers MutatorConnection object type with given service.
+func RegisterMutatorConnection(svc *graphql.Service, impl MutatorConnectionFieldResolvers) {
+	svc.RegisterObject(_ObjectTypeMutatorConnectionDesc, impl)
+}
+func _ObjTypeMutatorConnectionNodesHandler(impl interface{}) graphql1.FieldResolveFn {
+	resolver := impl.(MutatorConnectionNodesFieldResolver)
+	return func(frp graphql1.ResolveParams) (interface{}, error) {
+		return resolver.Nodes(frp)
+	}
+}
+
+func _ObjTypeMutatorConnectionPageInfoHandler(impl interface{}) graphql1.FieldResolveFn {
+	resolver := impl.(MutatorConnectionPageInfoFieldResolver)
+	return func(frp graphql1.ResolveParams) (interface{}, error) {
+		return resolver.PageInfo(frp)
+	}
+}
+
+func _ObjectTypeMutatorConnectionConfigFn() graphql1.ObjectConfig {
+	return graphql1.ObjectConfig{
+		Description: "A connection to a sequence of records.",
+		Fields: graphql1.Fields{
+			"nodes": &graphql1.Field{
+				Args:              graphql1.FieldConfigArgument{},
+				DeprecationReason: "",
+				Description:       "self descriptive",
+				Name:              "nodes",
+				Type:              graphql1.NewNonNull(graphql1.NewList(graphql1.NewNonNull(graphql.OutputType("Mutator")))),
+			},
+			"pageInfo": &graphql1.Field{
+				Args:              graphql1.FieldConfigArgument{},
+				DeprecationReason: "",
+				Description:       "self descriptive",
+				Name:              "pageInfo",
+				Type:              graphql1.NewNonNull(graphql.OutputType("OffsetPageInfo")),
+			},
+		},
+		Interfaces: []*graphql1.Interface{},
+		IsTypeOf: func(_ graphql1.IsTypeOfParams) bool {
+			// NOTE:
+			// Panic by default. Intent is that when Service is invoked, values of
+			// these fields are updated with instantiated resolvers. If these
+			// defaults are called it is most certainly programmer err.
+			// If you're see this comment then: 'Whoops! Sorry, my bad.'
+			panic("Unimplemented; see MutatorConnectionFieldResolvers.")
+		},
+		Name: "MutatorConnection",
+	}
+}
+
+// describe MutatorConnection's configuration; kept private to avoid unintentional tampering of configuration at runtime.
+var _ObjectTypeMutatorConnectionDesc = graphql.ObjectDesc{
+	Config: _ObjectTypeMutatorConnectionConfigFn,
+	FieldHandlers: map[string]graphql.FieldHandler{
+		"nodes":    _ObjTypeMutatorConnectionNodesHandler,
+		"pageInfo": _ObjTypeMutatorConnectionPageInfoHandler,
+	},
+}
+
+// MutatorEdgeNodeFieldResolver implement to resolve requests for the MutatorEdge's node field.
+type MutatorEdgeNodeFieldResolver interface {
+	// Node implements response to request for node field.
+	Node(p graphql.ResolveParams) (interface{}, error)
+}
+
+// MutatorEdgeCursorFieldResolver implement to resolve requests for the MutatorEdge's cursor field.
+type MutatorEdgeCursorFieldResolver interface {
+	// Cursor implements response to request for cursor field.
+	Cursor(p graphql.ResolveParams) (string, error)
+}
+
+//
+// MutatorEdgeFieldResolvers represents a collection of methods whose products represent the
+// response values of the 'MutatorEdge' type.
+//
+// == Example SDL
+//
+//   """
+//   Dog's are not hooman.
+//   """
+//   type Dog implements Pet {
+//     "name of this fine beast."
+//     name:  String!
+//
+//     "breed of this silly animal; probably shibe."
+//     breed: [Breed]
+//   }
+//
+// == Example generated interface
+//
+//   // DogResolver ...
+//   type DogFieldResolvers interface {
+//     DogNameFieldResolver
+//     DogBreedFieldResolver
+//
+//     // IsTypeOf is used to determine if a given value is associated with the Dog type
+//     IsTypeOf(interface{}, graphql.IsTypeOfParams) bool
+//   }
+//
+// == Example implementation ...
+//
+//   // DogResolver implements DogFieldResolvers interface
+//   type DogResolver struct {
+//     logger logrus.LogEntry
+//     store interface{
+//       store.BreedStore
+//       store.DogStore
+//     }
+//   }
+//
+//   // Name implements response to request for name field.
+//   func (r *DogResolver) Name(p graphql.ResolveParams) (interface{}, error) {
+//     // ... implementation details ...
+//     dog := p.Source.(DogGetter)
+//     return dog.GetName()
+//   }
+//
+//   // Breed implements response to request for breed field.
+//   func (r *DogResolver) Breed(p graphql.ResolveParams) (interface{}, error) {
+//     // ... implementation details ...
+//     dog := p.Source.(DogGetter)
+//     breed := r.store.GetBreed(dog.GetBreedName())
+//     return breed
+//   }
+//
+//   // IsTypeOf is used to determine if a given value is associated with the Dog type
+//   func (r *DogResolver) IsTypeOf(p graphql.IsTypeOfParams) bool {
+//     // ... implementation details ...
+//     _, ok := p.Value.(DogGetter)
+//     return ok
+//   }
+//
+type MutatorEdgeFieldResolvers interface {
+	MutatorEdgeNodeFieldResolver
+	MutatorEdgeCursorFieldResolver
+}
+
+// MutatorEdgeAliases implements all methods on MutatorEdgeFieldResolvers interface by using reflection to
+// match name of field to a field on the given value. Intent is reduce friction
+// of writing new resolvers by removing all the instances where you would simply
+// have the resolvers method return a field.
+//
+// == Example SDL
+//
+//    type Dog {
+//      name:   String!
+//      weight: Float!
+//      dob:    DateTime
+//      breed:  [Breed]
+//    }
+//
+// == Example generated aliases
+//
+//   type DogAliases struct {}
+//   func (_ DogAliases) Name(p graphql.ResolveParams) (interface{}, error) {
+//     // reflect...
+//   }
+//   func (_ DogAliases) Weight(p graphql.ResolveParams) (interface{}, error) {
+//     // reflect...
+//   }
+//   func (_ DogAliases) Dob(p graphql.ResolveParams) (interface{}, error) {
+//     // reflect...
+//   }
+//   func (_ DogAliases) Breed(p graphql.ResolveParams) (interface{}, error) {
+//     // reflect...
+//   }
+//
+// == Example Implementation
+//
+//   type DogResolver struct { // Implements DogResolver
+//     DogAliases
+//     store store.BreedStore
+//   }
+//
+//   // NOTE:
+//   // All other fields are satisified by DogAliases but since this one
+//   // requires hitting the store we implement it in our resolver.
+//   func (r *DogResolver) Breed(p graphql.ResolveParams) interface{} {
+//     dog := v.(*Dog)
+//     return r.BreedsById(dog.BreedIDs)
+//   }
+//
+type MutatorEdgeAliases struct{}
+
+// Node implements response to request for 'node' field.
+func (_ MutatorEdgeAliases) Node(p graphql.ResolveParams) (interface{}, error) {
+	val, err := graphql.DefaultResolver(p.Source, p.Info.FieldName)
+	return val, err
+}
+
+// Cursor implements response to request for 'cursor' field.
+func (_ MutatorEdgeAliases) Cursor(p graphql.ResolveParams) (string, error) {
+	val, err := graphql.DefaultResolver(p.Source, p.Info.FieldName)
+	ret, ok := val.(string)
+	if err != nil {
+		return ret, err
+	}
+	if !ok {
+		return ret, errors.New("unable to coerce value for field 'cursor'")
+	}
+	return ret, err
+}
+
+// MutatorEdgeType An edge in a connection.
+var MutatorEdgeType = graphql.NewType("MutatorEdge", graphql.ObjectKind)
+
+// RegisterMutatorEdge registers MutatorEdge object type with given service.
+func RegisterMutatorEdge(svc *graphql.Service, impl MutatorEdgeFieldResolvers) {
+	svc.RegisterObject(_ObjectTypeMutatorEdgeDesc, impl)
+}
+func _ObjTypeMutatorEdgeNodeHandler(impl interface{}) graphql1.FieldResolveFn {
+	resolver := impl.(MutatorEdgeNodeFieldResolver)
+	return func(frp graphql1.ResolveParams) (interface{}, error) {
+		return resolver.Node(frp)
+	}
+}
+
+func _ObjTypeMutatorEdgeCursorHandler(impl interface{}) graphql1.FieldResolveFn {
+	resolver := impl.(MutatorEdgeCursorFieldResolver)
+	return func(frp graphql1.ResolveParams) (interface{}, error) {
+		return resolver.Cursor(frp)
+	}
+}
+
+func _ObjectTypeMutatorEdgeConfigFn() graphql1.ObjectConfig {
+	return graphql1.ObjectConfig{
+		Description: "An edge in a connection.",
+		Fields: graphql1.Fields{
+			"cursor": &graphql1.Field{
+				Args:              graphql1.FieldConfigArgument{},
+				DeprecationReason: "",
+				Description:       "self descriptive",
+				Name:              "cursor",
+				Type:              graphql1.NewNonNull(graphql1.String),
+			},
+			"node": &graphql1.Field{
+				Args:              graphql1.FieldConfigArgument{},
+				DeprecationReason: "",
+				Description:       "self descriptive",
+				Name:              "node",
+				Type:              graphql.OutputType("Mutator"),
+			},
+		},
+		Interfaces: []*graphql1.Interface{},
+		IsTypeOf: func(_ graphql1.IsTypeOfParams) bool {
+			// NOTE:
+			// Panic by default. Intent is that when Service is invoked, values of
+			// these fields are updated with instantiated resolvers. If these
+			// defaults are called it is most certainly programmer err.
+			// If you're see this comment then: 'Whoops! Sorry, my bad.'
+			panic("Unimplemented; see MutatorEdgeFieldResolvers.")
+		},
+		Name: "MutatorEdge",
+	}
+}
+
+// describe MutatorEdge's configuration; kept private to avoid unintentional tampering of configuration at runtime.
+var _ObjectTypeMutatorEdgeDesc = graphql.ObjectDesc{
+	Config: _ObjectTypeMutatorEdgeConfigFn,
+	FieldHandlers: map[string]graphql.FieldHandler{
+		"cursor": _ObjTypeMutatorEdgeCursorHandler,
+		"node":   _ObjTypeMutatorEdgeNodeHandler,
+	},
+}
+
+// MutatorListOrder Describes ways in which a list of mutators can be ordered.
+type MutatorListOrder string
+
+// MutatorListOrders holds enum values
+var MutatorListOrders = _EnumTypeMutatorListOrderValues{
+	NAME:      "NAME",
+	NAME_DESC: "NAME_DESC",
+}
+
+// MutatorListOrderType Describes ways in which a list of mutators can be ordered.
+var MutatorListOrderType = graphql.NewType("MutatorListOrder", graphql.EnumKind)
+
+// RegisterMutatorListOrder registers MutatorListOrder object type with given service.
+func RegisterMutatorListOrder(svc *graphql.Service) {
+	svc.RegisterEnum(_EnumTypeMutatorListOrderDesc)
+}
+func _EnumTypeMutatorListOrderConfigFn() graphql1.EnumConfig {
+	return graphql1.EnumConfig{
+		Description: "Describes ways in which a list of mutators can be ordered.",
+		Name:        "MutatorListOrder",
+		Values: graphql1.EnumValueConfigMap{
+			"NAME": &graphql1.EnumValueConfig{
+				DeprecationReason: "",
+				Description:       "self descriptive",
+				Value:             "NAME",
+			},
+			"NAME_DESC": &graphql1.EnumValueConfig{
+				DeprecationReason: "",
+				Description:       "self descriptive",
+				Value:             "NAME_DESC",
+			},
+		},
+	}
+}
+
+// describe MutatorListOrder's configuration; kept private to avoid unintentional tampering of configuration at runtime.
+var _EnumTypeMutatorListOrderDesc = graphql.EnumDesc{Config: _EnumTypeMutatorListOrderConfigFn}
+
+type _EnumTypeMutatorListOrderValues struct {
+	// NAME - self descriptive
+	NAME MutatorListOrder
+	// NAME_DESC - self descriptive
+	NAME_DESC MutatorListOrder
+}