+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Silenced entries are now retrieved from the cache when determining if an event
 is silenced.
 - Added --disable-assets flag to sensu-agent.
+- Added ability to query mutators to the GraphQL service
 
 ### Fixed
 - Fixed a bug where events were not deleted when their corresponding entity was.